requirements/base.txt
# ═══════════════════════════════════════════════════════════════════
# Enterprise NGFW - Base Requirements
# ═══════════════════════════════════════════════════════════════════

# Web Framework & API
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.24.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.0
cryptography>=41.0.0
slowapi>=0.1.9

# Database ORM
SQLAlchemy>=2.0.0

# HTTP & Async
aiohttp>=3.8.0

# ML & Numerics
numpy>=1.24.0

# Monitoring & Metrics
prometheus-client>=0.17.0
psutil>=5.9.0

# Configuration
PyYAML>=6.0
jsonschema>=4.17.0

# Logging
python-json-logger>=2.0.0

# CLI
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Utilities
python-dateutil>=2.8.2

# eBPF Support (optional - requires BCC installation)
# python-bcc; platform_system=="Linux"

# Optional ML Runtimes (install as needed)
# onnxruntime>=1.16.0
# joblib>=1.3.0
# torch>=2.0.0
